refactored sfdr report tests
